feat(ldap)：添加注解，添加接口

+++ LDAP/src/main/java/cn/ldap/ldap/common/entity/SyncStatus.java
@@ -35,12 +35,12 @@ public class SyncStatus {
      * 主服务数量
      */
     @TableField("main_server_number")
-    private Integer mainServerNumber;
+    private Long mainServerNumber;
     /**
      * 从服务数量
      */
     @TableField("follow_server_number")
-    private Integer followServerNumber;
+    private Long followServerNumber;
     /**
      * 同步状态
      */
